refactor compression for float
