+++ tachyon_core/src/storage/compression/mod.rs
@@ -1,0 +1,24 @@
+use std::io::{Read, Write};
+
+use crate::Timestamp;
+
+use super::file::Header;
+
+pub mod int;
+
+pub trait CompressionEngine<W: Write> {
+    type PhysicalType;
+    fn new(writer: W, header: &Header) -> Self
+    where
+        Self: Sized;
+    fn consume(&mut self, timestamp: Timestamp, value: Self::PhysicalType);
+    fn flush_all(&mut self) -> usize;
+}
+
+pub trait DecompressionEngine<R: Read> {
+    type PhysicalType;
+    fn new(reader: R, header: &Header) -> Self
+    where
+        Self: Sized;
+    fn next(&mut self) -> (Timestamp, Self::PhysicalType);
+}
